+++ magicbell/api/__init__.py
@@ -1,4 +1,5 @@
 from .channels import ChannelsAPI  # noqa: F401
+from .graphql import GraphQLAPI  # noqa: F401
 from .project import ProjectAPI  # noqa: F401
 from .realtime import RealtimeAPI  # noqa: F401
 from .user import UserAPI  # noqa: F401